Number Guessing Game- My First C Project.
My first project made in C language wherein I used only if else statements to make the game work.
The machine will generate the random number between 1 to 100 and the user have to guess the number. if the input is close to the actual number, the machine will let the user know if the number is 5 more or less. in the end, if user guesses the number correctly, he/she will get the result that how many tries did it require to guess it.
Hope you like it, I would love to see your contribution if required!